Add VideoStrong KM5 (videostrong-km5)
Device specifications: https://whatismyphone.net/vendors/videostrong/km5-android-tv-box-km5

The manufacturer's website is here: https://www.videostrong.com/
(Unfortunately the original device page has been taken down)

This is a board that I've ripped the NAND module (usually has one) off
of it so I can't rely on the original bootloader tricks to get a newer version
on U-Boot running on it. Thankfully I was able to get it going by
dumping and extracting the necessary bootloader parts from the original
firmware dump I found somewhere and I've tested that it works.

The acs.bin dump is based on the ddrs_ tables found in bl2.bin, through
crude hexediting I was able to transfer that over to the Le Potato's
acs.bin, I've tested that and it seems to work. The rest of the files
are original from the ROM dump.